Query the grammar point overview from the DB
This commit also converts every DB id from an Int
to a Long. This is because the previous commit
did not properly insert reviews in DB. While
debugging this issue, I took the time to clean
the type of the DB classes.

In the end, multiple problems existed:
- the sync was successful even though the reviews
  were not inserted in DB
- two reviews diff were applied, which resulted
  in only the ghost reviews being inserted and
  the history points not being inserted because
  of a foreign key constraint violation

+++ app/src/main/java/dev/esnault/bunpyro/android/screen/grammarpoint/GrammarPointViewModel.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.withContext
 
 
 class GrammarPointViewModel(
-    id: Int,
+    id: Long,
     private val grammarRepo: IGrammarPointRepository
 ) : BaseViewModel() {
 
@@ -30,7 +30,7 @@ class GrammarPointViewModel(
         loadGrammarPoint(id)
     }
 
-    private fun loadGrammarPoint(id: Int) {
+    private fun loadGrammarPoint(id: Long) {
         viewModelScope.launch {
             withContext(Dispatchers.IO) {
                 // TODO Handle the errors
@@ -59,7 +59,7 @@ class GrammarPointViewModel(
         this.currentState = currentState.copy(furiganaShown = !currentState.furiganaShown)
     }
 
-    fun onGrammarPointClick(id: Int) {
+    fun onGrammarPointClick(id: Long) {
         navigate(GrammarPointFragmentDirections.actionGrammarPointToGrammarPoint(id))
     }
