+++ app/src/main/java/dev/esnault/bunpyro/data/db/examplesentence/ExampleSentenceDb.kt
@@ -20,8 +20,8 @@ import dev.esnault.bunpyro.data.db.grammarpoint.GrammarPointDb
     ]
 )
 data class ExampleSentenceDb(
-    @PrimaryKey val id: Int,
-    @ColumnInfo(name = "grammar_id", index = true) val grammarId: Int,
+    @PrimaryKey val id: Long,
+    @ColumnInfo(name = "grammar_id", index = true) val grammarId: Long,
     val japanese: String,
     val english: String,
     val nuance: String?,
