+++ app/src/main/java/dev/esnault/bunpyro/data/db/review/ReviewDb.kt
@@ -8,9 +8,9 @@ import java.util.*
 
 @Entity(tableName = "review")
 data class ReviewDb(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     val type: ReviewType,
-    @ColumnInfo(name = "grammar_id") val grammarId: Int,
+    @ColumnInfo(name = "grammar_id") val grammarId: Long,
     @ColumnInfo(name = "created_at") val createdAt: Date,
     @ColumnInfo(name = "updated_at") val updatedAt: Date,
     @ColumnInfo(name = "next_review") val nextReview: Date,
