+++ app/src/main/java/dev/esnault/bunpyro/data/network/entities/ExampleSentence.kt
@@ -3,12 +3,12 @@ package dev.esnault.bunpyro.data.network.entities
 import com.squareup.moshi.Json
 
 data class ExampleSentence(
-    val id: Int,
+    val id: Long,
     val attributes: Attributes
 ) {
 
     data class Attributes(
-        @Json(name = "grammar-point-id") val grammarId: Int,
+        @Json(name = "grammar-point-id") val grammarId: Long,
         val japanese: String,
         val english: String,
         val nuance: String?,
