+++ app/src/main/java/dev/esnault/bunpyro/data/network/entities/NormalReview.kt
@@ -25,15 +25,15 @@ import java.util.*
  * - review_misses
  */
 data class NormalReview(
-    val id: Int,
-    @Json(name = "study_question_id") val questionId: Int,
-    @Json(name = "grammar_point_id") val grammarId: Int,
+    val id: Long,
+    @Json(name = "study_question_id") val questionId: Long,
+    @Json(name = "grammar_point_id") val grammarId: Long,
     @Json(name = "next_review") val nextReview: Date,
     @Json(name = "created_at") val createdAt: Date,
     @Json(name = "updated_at") val updatedAt: Date,
     @Json(name = "last_studied_at") val lastStudiedAt: Date?,
-    val readings: List<Int>,
+    val readings: List<Long>,
     val history: List<ReviewHistory>,
-    @Json(name = "missed_question_ids") val missedQuestionIds: List<Int>,
-    @Json(name = "studied_question_ids") val studiedQuestionIds: List<Int>
+    @Json(name = "missed_question_ids") val missedQuestionIds: List<Long>,
+    @Json(name = "studied_question_ids") val studiedQuestionIds: List<Long>
 )
