+++ app/src/main/java/dev/esnault/bunpyro/data/network/entities/SupplementalLink.kt
@@ -4,12 +4,12 @@ import com.squareup.moshi.Json
 
 
 data class SupplementalLink(
-    val id: Int,
+    val id: Long,
     val attributes: Attributes
 ) {
 
     data class Attributes(
-        @Json(name = "grammar-point-id") val grammarId: Int,
+        @Json(name = "grammar-point-id") val grammarId: Long,
         val site: String,
         val link: String,
         val description: String
