+++ app/src/main/java/dev/esnault/bunpyro/data/repository/grammarpoint/IGrammarPointRepository.kt
@@ -5,5 +5,5 @@ import dev.esnault.bunpyro.domain.entities.grammar.GrammarPoint
 
 interface IGrammarPointRepository {
 
-    suspend fun getGrammarPoint(id: Int): GrammarPoint
+    suspend fun getGrammarPoint(id: Long): GrammarPoint
 }
